requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
cryptography==43.0.0
fastapi==0.124.2
google-api-python-client==2.154.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
h11==0.16.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
opencc-python-reimplemented==0.1.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
soupsieve==2.8
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
